+++ pirel/__init__.py
@@ -4,7 +4,7 @@ from rich.console import Console
 
 from pirel.releases import PythonReleases, load_releases
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
 
 
 class PirelContext: