+++ CHANGELOG.md
@@ -19,6 +19,9 @@ Types of changes:
 
 ## [Unreleased]
 
+
+## [0.3.0] - 2025-01-04
+
 ### Added
 * Add global option `--version` (#10)
 * Create cache of release cycle data and add option `--no-cache` to clear cache (#12)
@@ -77,7 +80,8 @@ by answering questions based on the release cycle data (#13)
 * Basic CLI app that shows all Python releases with the active Python interpeter being highlighted
 
 
-[unreleased]: https://github.com/RafaelWO/pirel/compare/0.2.1...HEAD
+[unreleased]: https://github.com/RafaelWO/pirel/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/RafaelWO/pirel/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/RafaelWO/pirel/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/RafaelWO/pirel/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/RafaelWO/pirel/compare/0.1.0...0.1.1